+++ example/http/server/flex/http_server_flex.cpp
@@ -270,7 +270,7 @@ class session
         }
     };
 
-    std::string const& doc_root_;
+    std::shared_ptr<std::string const> doc_root_;
     http::request<http::string_body> req_;
     std::shared_ptr<void> res_;
     send_lambda lambda_;
@@ -286,7 +286,7 @@ public:
     session(
         boost::asio::io_context& ioc,
         boost::beast::flat_buffer buffer,
-        std::string const& doc_root)
+        std::shared_ptr<std::string const> const& doc_root)
         : doc_root_(doc_root)
         , lambda_(*this)
         , strand_(ioc.get_executor())
@@ -326,7 +326,7 @@ public:
             return fail(ec, "read");
 
         // Send the response
-        handle_request(doc_root_, std::move(req_), lambda_);
+        handle_request(*doc_root_, std::move(req_), lambda_);
     }
 
     void
@@ -369,7 +369,7 @@ public:
     plain_session(
         tcp::socket socket,
         boost::beast::flat_buffer buffer,
-        std::string const& doc_root)
+        std::shared_ptr<std::string const> const& doc_root)
         : session<plain_session>(
             socket.get_executor().context(),
             std::move(buffer),
@@ -420,7 +420,7 @@ public:
         tcp::socket socket,
         ssl::context& ctx,
         boost::beast::flat_buffer buffer,
-        std::string const& doc_root)
+        std::shared_ptr<std::string const> const& doc_root)
         : session<ssl_session>(
             socket.get_executor().context(),
             std::move(buffer),
@@ -501,7 +501,7 @@ class detect_session : public std::enable_shared_from_this<detect_session>
     ssl::context& ctx_;
     boost::asio::strand<
         boost::asio::io_context::executor_type> strand_;
-    std::string const& doc_root_;
+    std::shared_ptr<std::string const> doc_root_;
     boost::beast::flat_buffer buffer_;
 
 public:
@@ -509,7 +509,7 @@ public:
     detect_session(
         tcp::socket socket,
         ssl::context& ctx,
-        std::string const& doc_root)
+        std::shared_ptr<std::string const> const& doc_root)
         : socket_(std::move(socket))
         , ctx_(ctx)
         , strand_(socket_.get_executor())
@@ -567,14 +567,14 @@ class listener : public std::enable_shared_from_this<listener>
         boost::asio::io_context::executor_type> strand_;
     tcp::acceptor acceptor_;
     tcp::socket socket_;
-    std::string const& doc_root_;
+    std::shared_ptr<std::string const> doc_root_;
 
 public:
     listener(
         boost::asio::io_context& ioc,
         ssl::context& ctx,
         tcp::endpoint endpoint,
-        std::string const& doc_root)
+        std::shared_ptr<std::string const> const& doc_root)
         : ctx_(ctx)
         , strand_(ioc.get_executor())
         , acceptor_(ioc)
@@ -673,7 +673,7 @@ int main(int argc, char* argv[])
     }
     auto const address = boost::asio::ip::make_address(argv[1]);
     auto const port = static_cast<unsigned short>(std::atoi(argv[2]));
-    std::string const doc_root = argv[3];
+    auto const doc_root = std::make_shared<std::string>(argv[3]);
     auto const threads = std::max<int>(1, std::atoi(argv[4]));
 
     // The io_context is required for all I/O